chore: group cognito types
chore: add docstring to AuthFlowType

chore: export CognitoResendSignUpCodeOptions

chore: export ResendSignUpCode Options

chore: export cognito user attribute key

chore: export cognito user attribute key
