+++ packages/auth/src/providers/cognito/apis/resendSignUpCode.ts
@@ -13,8 +13,8 @@ import {
 import { assertValidationError } from '../../../errors/utils/assertValidationError';
 import { AuthValidationErrorCode } from '../../../errors/types/validation';
 import { resendSignUpConfirmationCodeClient } from '../utils/clients/ResendSignUpCodeClient';
-import type { CognitoResendSignUpCodeOptions } from '../types/options/CognitoResendSignUpCodeOptions';
-import type { CognitoUserAttributeKey } from '../types/models/CognitoUserAttributeKey';
+import { CognitoResendSignUpCodeOptions } from '../types';
+
 
 /**
  * Resend the confirmation code while signing up
