+++ packages/auth/src/providers/cognito/types/errors.ts
