+++ packages/auth/src/providers/cognito/types/index.ts
@@ -1,8 +1,15 @@
-export { ClientMetadata } from './models/ClientMetadata';
-export { CustomAttribute } from './models/CustomAttribute';
-export { ValidationData } from './models/ValidationData';
-export { AuthFlowType } from './models/AuthFlowType';
-export { CognitoConfirmResetPasswordOptions } from './options/CognitoConfirmResetPasswordOptions';
-export { CognitoSignUpOptions } from './options/CognitoSignUpOptions';
-export { CognitoResetPasswordOptions } from './options/CognitoResetPasswordOptions';
-export { CognitoSignInOptions } from './options/CognitoSignInOptions';
+export {
+	ClientMetadata,
+	CustomAttribute,
+	ValidationData,
+	AuthFlowType,
+    CognitoUserAttributeKey
+} from './models';
+
+export {
+	CognitoConfirmResetPasswordOptions,
+	CognitoSignUpOptions,
+	CognitoResetPasswordOptions,
+	CognitoSignInOptions,
+	CognitoResendSignUpCodeOptions,
+} from './options';
