--- packages/auth/src/providers/cognito/types/models/ClientMetadata.ts
@@ -1,9 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/**
- * Arbitrary key/value pairs that may be passed as part of certain Cognito requests
- */
-export type ClientMetadata = {
-	[key: string]: string;
-};
