--- packages/auth/src/providers/cognito/types/models/CustomAttribute.ts
@@ -1,7 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/**
- * Cognito custom attribute type
- */
-export type CustomAttribute = { custom: string };
