--- packages/auth/src/providers/cognito/types/models/ValidationData.ts
@@ -1,7 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-/**
- * One or more name-value pairs containing the validation data in the request to register a user.
- */
-export type ValidationData = { [key: string]: string };
