--- packages/auth/src/providers/cognito/types/options/CognitoResetPasswordOptions.ts
@@ -1,8 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-import { ClientMetadata } from '../models/ClientMetadata';
-
-export type CognitoResetPasswordOptions = {
-	clientMetadata?: ClientMetadata;
-};
